move rpi-specific infos to gpio-rpi.cc - create emums.h - add SetTxPower method to GPIO class
